code re-organization: moved health event-loop to a new file, moved health globals to health.c
