+++ CMakeLists.txt
@@ -899,6 +899,7 @@ set(HEALTH_PLUGIN_FILES
         health/health_silencers.h
         health/health_internals.h
         health/health_notifications.c
+        health/health_event_loop.c
 )
 
 set(IDLEJITTER_PLUGIN_FILES collectors/idlejitter.plugin/plugin_idlejitter.c)
